// src/figma-plugin/src/types.ts
export interface BaseNodeParams {
  nodeId: string;
}

export interface BasePositionParams {
  x: number;
  y: number;
}

export interface BaseSizeParams {
  width: number;
  height: number;
}
// ...
export interface MoveNodeParams extends BaseNodeParams, BasePositionParams {}

export interface ResizeNodeParams extends BaseNodeParams, BaseSizeParams {}

export interface CloneNodeParams extends BaseNodeParams {
  x?: number;
  y?: number;
}

export interface DeleteNodeParams extends BaseNodeParams {}

export interface DeleteMultipleNodesParams {
  nodeIds: string[];
}
// ...
export interface NodeInfo {
  id: string;
  name: string;
  type: string;
  visible?: boolean;
  x?: number;
  y?: number;
  width?: number;
  height?: number;
}